Reduce warning/suggestion count to zero.

+++ osu.Game/Screens/GameScreenWhiteBox.cs
@@ -21,7 +21,7 @@ namespace osu.Game.Screens
     {
         private BackButton popButton;
 
-        private const int transition_time = 1000;
+        private const double transition_time = 1000;
 
         protected virtual IEnumerable<Type> PossibleChildren => null;
